# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PreventivoAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business-profile:
    get:
      operationId: getBusinessProfile
      tags: [business-profile]
      summary: Get the authenticated user's business profile
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
    put:
      operationId: updateBusinessProfile
      tags: [business-profile]
      summary: Create or update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessProfileBody"
      responses:
        "200":
          description: Updated business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
  /business-profile/logo:
    post:
      operationId: uploadBusinessProfileLogo
      tags: [business-profile]
      summary: Upload or replace the company logo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                logo:
                  type: string
                  format: binary
              required:
                - logo
      responses:
        "200":
          description: Logo uploaded and profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoUploadResult"
...
components:
  schemas:
...
    BusinessProfile:
      type: object
      properties:
        userId:
          type: string
        companyName:
          type: string
        vatNumber:
          type: string
        address:
          type: string
        logoUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        apiKey:
          type: string
          nullable: true
      required:
        - userId
        - companyName
    UpdateBusinessProfileBody:
      type: object
      properties:
        companyName:
          type: string
        vatNumber:
          type: string
        address:
          type: string
        logoUrl:
          type: string
        phone:
          type: string
        email:
          type: string
...
    LogoUploadResult:
      type: object
      properties:
        logoUrl:
          type: string
      required:
        - logoUrl
...